backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
google-generativeai==0.7.2
requests==2.32.3
beautifulsoup4==4.12.3
rapidfuzz==3.9.6
numpy==2.1.1
scikit-learn==1.5.2
python-dotenv==1.0.1
pypdf==5.0.0
SQLAlchemy==2.0.34
sqlalchemy

